Fix CD of the Week multiple review display
This PR fixes an issue where multiple CD of the Week reviews with the same date weren't all displaying on the main cdoftheweek.php page. Previously, when two CDOTW reviews had the same date, they would both be displayed, but recently only one was showing.

The changes ensure that:
1. All CD reviews with the most recent date are shown on the main page
2. Date is displayed once at the top
3. Each review is displayed with proper artist, title, label, and reviewer info
4. Multiple reviews are separated with a horizontal rule

This restores the previous functionality that allowed multiple reviews to be featured simultaneously.